% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.grpss.R
\name{print.grpss}
\alias{print.grpss}
\title{Print the \code{grpss} object}
\usage{
\method{print}{grpss}(x, ...)
}
\arguments{
\item{x}{A fitted "\code{grpss}" object.}

\item{...}{Not used.}
}
\value{
The \code{call}, \code{criterion}, \code{threshold} and \code{screened groups} will
be printed out.
}
\description{
Prints the "\code{grpss}" object obtained from \code{\link{grpss}}.
}
\details{
This function prints out the group screening results obtained from
\code{\link{grpss}} when "\code{selection = FALSE}".
}
\author{
Debin Qiu, Jeongyoun Ahn
}
\seealso{
\code{\link{grpss}}
}

